HPCC-26323 Add option to persist Syntax Check WorkUnits

+++ esp/services/ws_workunits/ws_workunitsService.cpp
@@ -1285,7 +1285,8 @@ bool CWsWorkunitsEx::onWUSyntaxCheckECL(IEspContext &context, IEspWUSyntaxCheckR
         case WUStateAborted:
         case WUStateCompleted:
         case WUStateFailed:
-            factory->deleteWorkUnitEx(wuid.str(), true);
+            if (!req.getPersistWorkunit())
+                factory->deleteWorkUnitEx(wuid.str(), true);
             break;
         default:
             abortWorkUnit(wuid.str(), context.querySecManager(), context.queryUser());
